order_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
prometheus-fastapi-instrumentator>=5.0.0
prometheus-client>=0.14.0